Moved message related structs from offst-funder to offst-proto at funder/messages.rs
