+++ components/funder/src/lib.rs
@@ -7,20 +7,9 @@
 #![feature(dbg_macro)]
 #![cfg_attr(not(feature = "cargo-clippy"), allow(unknown_lints))]
 
-// TODO: Remove this later:
-
-// extern crate byteorder;
-// extern crate bytes;
-// extern crate futures;
-// extern crate futures_await as futures;
-// extern crate futures;
 extern crate futures_cpupool;
 #[macro_use]
 extern crate log;
-// extern crate rusqlite;
-// extern crate tokio_core;
-// extern crate tokio_io;
-// extern crate tokio_codec;
 
 extern crate num_traits;
 extern crate num_bigint;
@@ -28,17 +17,8 @@ extern crate num_bigint;
 extern crate serde;
 #[macro_use]
 extern crate serde_derive;
-// extern crate serde_json;
-// extern crate base64;
 
-// extern crate atomicwrites;
-
-// extern crate im;
-
-#[macro_use]
 extern crate utils;
-// extern crate offst_crypto as crypto;
-// extern crate offst_identity as identity;
 
 
 mod liveness;
