+++ components/funder/src/mutual_credit/outgoing.rs
@@ -9,14 +9,16 @@ use crypto::hash;
 use utils::safe_arithmetic::SafeSignedArithmetic;
 use utils::int_convert::usize_to_u32;
 
+use proto::funder::messages::{FriendTcOp, RequestSendFunds, 
+    ResponseSendFunds, FailureSendFunds,FriendsRoute,
+    InvoiceId};
+
 use super::types::{MutualCredit, McMutation, 
     MAX_FUNDER_DEBT};
-use super::super::credit_calc::CreditCalculator;
-use super::super::types::{FriendTcOp, RequestSendFunds, 
-    ResponseSendFunds, FailureSendFunds,
-    PendingFriendRequest, FriendsRoute, RequestsStatus,
-    SendFundsReceipt, InvoiceId};
-use super::super::signature_buff::{create_response_signature_buffer, 
+use crate::credit_calc::CreditCalculator;
+use crate::types::{PendingFriendRequest, RequestsStatus,
+    SendFundsReceipt, create_pending_request};
+use crate::signature_buff::{create_response_signature_buffer, 
     verify_failure_signature};
 
 
@@ -226,7 +228,7 @@ impl OutgoingMc {
         }
 
         // Add pending request funds:
-        let pending_friend_request = request_send_funds.create_pending_request();
+        let pending_friend_request = create_pending_request(&request_send_funds);
 
         let mut tc_mutations = Vec::new();
         let tc_mutation = McMutation::InsertLocalPendingRequest(pending_friend_request);
