+++ components/funder/src/mutual_credit/tests.rs
@@ -3,13 +3,15 @@ use crypto::test_utils::DummyRandom;
 use crypto::uid::{Uid, UID_LEN};
 use crypto::identity::{Identity, SoftwareEd25519Identity,
                         generate_pkcs8_key_pair, SIGNATURE_LEN, Signature};
-#[allow(unused)]
+
 use crypto::crypto_rand::{RandValue, RAND_VALUE_LEN};
+
+use proto::funder::messages::{InvoiceId, INVOICE_ID_LEN, FreezeLink, Ratio, FriendsRoute, 
+    RequestSendFunds, ResponseSendFunds, FailureSendFunds, FriendTcOp};
+
 use crate::mutual_credit::types::MutualCredit;
 #[allow(unused)]
-use crate::types::{RequestsStatus, InvoiceId, INVOICE_ID_LEN, 
-    FreezeLink, Ratio, FriendsRoute, 
-    RequestSendFunds, ResponseSendFunds, FailureSendFunds, FriendTcOp};
+use crate::types::{RequestsStatus, create_pending_request};
 use crate::signature_buff::{create_response_signature_buffer, 
     create_failure_signature_buffer};
 
@@ -117,7 +119,7 @@ fn test_request_response_send_funds() {
         freeze_links: vec![funder_freeze_link],
     };
 
-    let pending_request = request_send_funds.create_pending_request();
+    let pending_request = create_pending_request(&request_send_funds);
     apply_outgoing(&mut mutual_credit, FriendTcOp::RequestSendFunds(request_send_funds)).unwrap();
 
     assert_eq!(mutual_credit.state().balance.balance, 0);
@@ -194,7 +196,7 @@ fn test_request_failure_send_funds() {
         freeze_links: vec![funder_freeze_link],
     };
 
-    let pending_request = request_send_funds.create_pending_request();
+    let pending_request = create_pending_request(&request_send_funds);
     apply_outgoing(&mut mutual_credit, FriendTcOp::RequestSendFunds(request_send_funds)).unwrap();
 
     assert_eq!(mutual_credit.state().balance.balance, 0);
