+++ components/funder/src/signature_buff.rs
@@ -3,8 +3,10 @@
 use byteorder::{BigEndian, WriteBytesExt};
 use crypto::hash;
 use crypto::identity::{verify_signature, PublicKey};
-use super::types::{ResponseSendFunds, FailureSendFunds, 
-    PendingFriendRequest, SendFundsReceipt};
+
+use proto::funder::messages::{ResponseSendFunds, FailureSendFunds};
+use crate::types::{PendingFriendRequest, SendFundsReceipt};
+
 
 pub const FUND_SUCCESS_PREFIX: &[u8] = b"FUND_SUCCESS";
 pub const FUND_FAILURE_PREFIX: &[u8] = b"FUND_FAILURE";
