+++ components/funder/src/state.rs
@@ -7,10 +7,12 @@ use num_traits::identities::Zero;
 use crypto::identity::PublicKey;
 use crypto::uid::Uid;
 
+use proto::funder::messages::Ratio;
+
 use utils::int_convert::usize_to_u64;
 
-use super::friend::{FriendState, FriendMutation};
-use super::types::{SendFundsReceipt, Ratio, AddFriend};
+use crate::friend::{FriendState, FriendMutation};
+use crate::types::{SendFundsReceipt, AddFriend};
 
 
 #[derive(Clone, Serialize, Deserialize)]
