+++ components/funder/src/types.rs
@@ -1,5 +1,3 @@
-use std::collections::hash_set::HashSet;
-
 use byteorder::{WriteBytesExt, BigEndian};
 
 use utils::int_convert::{usize_to_u64};
@@ -7,21 +5,20 @@ use utils::int_convert::{usize_to_u64};
 use crypto::identity::{PublicKey, Signature, verify_signature};
 use crypto::uid::Uid;
 use crypto::crypto_rand::RandValue;
-use crypto::hash;
 use crypto::hash::{HashResult, sha_512_256};
 
+use proto::funder::messages::{FriendsRoute, InvoiceId, 
+    RequestSendFunds, FriendMoveToken, FriendMessage,
+    FriendTcOp};
+
 use identity::IdentityClient;
+
 use super::report::{FunderReport, FunderReportMutation};
 
 // Prefix used for chain hashing of token channel funds.
 // NEXT is used for hashing for the next move token funds.
 const TOKEN_NEXT: &[u8] = b"NEXT";
 
-
-pub const INVOICE_ID_LEN: usize = 32;
-
-/// The universal unique identifier of an invoice.
-define_fixed_bytes!(InvoiceId, INVOICE_ID_LEN);
 
 // pub const CHANNEL_TOKEN_LEN: usize = 32;
 // define_fixed_bytes!(ChannelToken, CHANNEL_TOKEN_LEN);
@@ -53,15 +50,6 @@ impl RequestsStatus {
 }
 
 
-#[derive(Eq, PartialEq, Debug, Clone, Serialize, Deserialize)]
-pub enum FriendTcOp {
-    EnableRequests,
-    DisableRequests,
-    SetRemoteMaxDebt(u128),
-    RequestSendFunds(RequestSendFunds),
-    ResponseSendFunds(ResponseSendFunds),
-    FailureSendFunds(FailureSendFunds),
-}
 
 #[derive(Eq, PartialEq, Debug, Clone, Serialize, Deserialize)]
 pub struct PendingFriendRequest {
@@ -71,27 +59,15 @@ pub struct PendingFriendRequest {
     pub invoice_id: InvoiceId,
 }
 
-
-#[derive(Eq, PartialEq, Debug, Clone, Serialize, Deserialize)]
-pub struct ResponseSendFunds {
-    pub request_id: Uid,
-    pub rand_nonce: RandValue,
-    pub signature: Signature,
+pub fn create_pending_request(request_send_funds: &RequestSendFunds) -> PendingFriendRequest {
+    PendingFriendRequest {
+        request_id: request_send_funds.request_id,
+        route: request_send_funds.route.clone(),
+        dest_payment: request_send_funds.dest_payment,
+        invoice_id: request_send_funds.invoice_id.clone(),
+    }
 }
 
-
-/// The ratio can be numeration / T::max_value(), or 1
-#[derive(PartialEq, Eq, Debug, Clone, Serialize, Deserialize)]
-pub enum Ratio<T> {
-    One,
-    Numerator(T),
-}
-
-#[derive(Eq, PartialEq, Debug, Clone, Serialize, Deserialize)]
-pub struct FreezeLink {
-    pub shared_credits: u128,
-    pub usable_ratio: Ratio<u128>
-}
 
 /// A request to send funds that originates from the user
 #[derive(Debug, Clone)]
@@ -102,45 +78,6 @@ pub struct UserRequestSendFunds {
     pub dest_payment: u128,
 }
 
-
-#[derive(Eq, PartialEq, Debug, Clone, Serialize, Deserialize)]
-pub struct RequestSendFunds {
-    pub request_id: Uid,
-    pub route: FriendsRoute,
-    pub dest_payment: u128,
-    pub invoice_id: InvoiceId,
-    pub freeze_links: Vec<FreezeLink>,
-}
-
-
-#[derive(Eq, PartialEq, Debug, Clone, Serialize, Deserialize)]
-pub struct FailureSendFunds {
-    pub request_id: Uid,
-    pub reporting_public_key: PublicKey,
-    pub rand_nonce: RandValue,
-    pub signature: Signature,
-}
-
-
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-pub struct FriendsRoute {
-    pub public_keys: Vec<PublicKey>,
-}
-
-
-#[allow(unused)]
-#[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize)]
-pub struct FriendMoveToken {
-    pub operations: Vec<FriendTcOp>,
-    pub old_token: Signature,
-    pub inconsistency_counter: u64,
-    pub move_token_counter: u128,
-    pub balance: i128,
-    pub local_pending_debt: u128,
-    pub remote_pending_debt: u128,
-    pub rand_nonce: RandValue,
-    pub new_token: Signature,
-}
 
 #[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize)]
 pub struct FriendMoveTokenHashed {
@@ -156,8 +93,7 @@ pub struct FriendMoveTokenHashed {
 }
 
 
-impl FriendMoveToken {
-    pub async fn new(operations: Vec<FriendTcOp>,
+pub async fn create_friend_move_token(operations: Vec<FriendTcOp>,
                  old_token: Signature,
                  inconsistency_counter: u64,
                  move_token_counter: u128,
@@ -167,265 +103,74 @@ impl FriendMoveToken {
                  rand_nonce: RandValue,
                  identity_client: IdentityClient) -> FriendMoveToken {
 
-        let mut friend_move_token = FriendMoveToken {
-            operations,
-            old_token,
-            inconsistency_counter,
-            move_token_counter,
-            balance,
-            local_pending_debt,
-            remote_pending_debt,
-            rand_nonce,
-            new_token: Signature::zero(),
-        };
+    let mut friend_move_token = FriendMoveToken {
+        operations,
+        old_token,
+        inconsistency_counter,
+        move_token_counter,
+        balance,
+        local_pending_debt,
+        remote_pending_debt,
+        rand_nonce,
+        new_token: Signature::zero(),
+    };
 
-        let sig_buffer = friend_move_token.signature_buff();
-        friend_move_token.new_token = await!(identity_client.request_signature(sig_buffer)).unwrap();
-        friend_move_token
+    let sig_buffer = signature_buff(&friend_move_token);
+    friend_move_token.new_token = await!(identity_client.request_signature(sig_buffer)).unwrap();
+    friend_move_token
+}
+
+/// Combine all operations into one hash value.
+fn operations_hash(friend_move_token: &FriendMoveToken) -> HashResult {
+    let mut operations_data = Vec::new();
+    operations_data.write_u64::<BigEndian>(
+        usize_to_u64(friend_move_token.operations.len()).unwrap()).unwrap();
+    for op in &friend_move_token.operations {
+        operations_data.extend_from_slice(&op.to_bytes());
     }
+    sha_512_256(&operations_data)
+}
 
-    /// Combine all operations into one hash value.
-    fn operations_hash(&self) -> HashResult {
-        let mut operations_data = Vec::new();
-        operations_data.write_u64::<BigEndian>(
-            usize_to_u64(self.operations.len()).unwrap()).unwrap();
-        for op in &self.operations {
-            operations_data.extend_from_slice(&op.to_bytes());
-        }
-        sha_512_256(&operations_data)
-    }
+fn signature_buff(friend_move_token: &FriendMoveToken) -> Vec<u8> {
+    let mut sig_buffer = Vec::new();
+    sig_buffer.extend_from_slice(&sha_512_256(TOKEN_NEXT));
+    sig_buffer.extend_from_slice(&operations_hash(friend_move_token));
+    sig_buffer.extend_from_slice(&friend_move_token.old_token);
+    sig_buffer.write_u64::<BigEndian>(friend_move_token.inconsistency_counter).unwrap();
+    sig_buffer.write_u128::<BigEndian>(friend_move_token.move_token_counter).unwrap();
+    sig_buffer.write_i128::<BigEndian>(friend_move_token.balance).unwrap();
+    sig_buffer.write_u128::<BigEndian>(friend_move_token.local_pending_debt).unwrap();
+    sig_buffer.write_u128::<BigEndian>(friend_move_token.remote_pending_debt).unwrap();
+    sig_buffer.extend_from_slice(&friend_move_token.rand_nonce);
 
-    fn signature_buff(&self) -> Vec<u8> {
-        let mut sig_buffer = Vec::new();
-        sig_buffer.extend_from_slice(&sha_512_256(TOKEN_NEXT));
-        sig_buffer.extend_from_slice(&self.operations_hash());
-        sig_buffer.extend_from_slice(&self.old_token);
-        sig_buffer.write_u64::<BigEndian>(self.inconsistency_counter).unwrap();
-        sig_buffer.write_u128::<BigEndian>(self.move_token_counter).unwrap();
-        sig_buffer.write_i128::<BigEndian>(self.balance).unwrap();
-        sig_buffer.write_u128::<BigEndian>(self.local_pending_debt).unwrap();
-        sig_buffer.write_u128::<BigEndian>(self.remote_pending_debt).unwrap();
-        sig_buffer.extend_from_slice(&self.rand_nonce);
+    sig_buffer
+}
 
-        sig_buffer
-    }
+/// Verify that new_token is a valid signature over the rest of the fields.
+pub fn verify_friend_move_token(friend_move_token: &FriendMoveToken, public_key: &PublicKey) -> bool {
+    let sig_buffer = signature_buff(friend_move_token);
+    verify_signature(&sig_buffer, public_key, &friend_move_token.new_token)
+}
 
-    /// Verify that new_token is a valid signature over the rest of the fields.
-    pub fn verify(&self, public_key: &PublicKey) -> bool {
-        let sig_buffer = self.signature_buff();
-        verify_signature(&sig_buffer, public_key, &self.new_token)
-    }
-
-    /// Create a hashed version of the FriendMoveToken.
-    /// Hashed version contains the hash of the operations instead of the operations themselves,
-    /// hence it is usually shorter.
-    pub fn create_hashed(&self) -> FriendMoveTokenHashed {
-        FriendMoveTokenHashed {
-            operations_hash: self.operations_hash(),
-            old_token: self.old_token.clone(),
-            inconsistency_counter: self.inconsistency_counter,
-            move_token_counter: self.move_token_counter,
-            balance: self.balance,
-            local_pending_debt: self.local_pending_debt,
-            remote_pending_debt: self.remote_pending_debt,
-            rand_nonce: self.rand_nonce.clone(),
-            new_token: self.new_token.clone(),
-        }
+/// Create a hashed version of the FriendMoveToken.
+/// Hashed version contains the hash of the operations instead of the operations themselves,
+/// hence it is usually shorter.
+pub fn create_hashed(friend_move_token: &FriendMoveToken) -> FriendMoveTokenHashed {
+    FriendMoveTokenHashed {
+        operations_hash: operations_hash(friend_move_token),
+        old_token: friend_move_token.old_token.clone(),
+        inconsistency_counter: friend_move_token.inconsistency_counter,
+        move_token_counter: friend_move_token.move_token_counter,
+        balance: friend_move_token.balance,
+        local_pending_debt: friend_move_token.local_pending_debt,
+        remote_pending_debt: friend_move_token.remote_pending_debt,
+        rand_nonce: friend_move_token.rand_nonce.clone(),
+        new_token: friend_move_token.new_token.clone(),
     }
 }
 
 
-impl FriendsRoute {
-    /*
-    pub fn bytes_count(&self) -> usize {
-        mem::size_of::<PublicKey>() * 2 +
-            FriendRouteLink::bytes_count()
-                * self.route_links.len()
-    }
-    */
 
-    pub fn len(&self) -> usize {
-        self.public_keys.len()
-    }
-
-    /// Check if the route is valid.
-    /// A valid route must have at least 2 nodes, and is in one of the following forms:
-    /// A -- B -- C -- D -- E -- F -- A   (Single cycle, first == last)
-    /// A -- B -- C -- D -- E -- F        (A route with no repetitions)
-    pub fn is_valid(&self) -> bool {
-        if self.public_keys.len() < 2 {
-            return false;
-        }
-
-        let mut seen = HashSet::new();
-        for public_key in &self.public_keys[.. self.public_keys.len() - 1] {
-            if !seen.insert(public_key.clone()) {
-                return false;
-            }
-        }
-        let last_pk = &self.public_keys[self.public_keys.len() - 1];
-        if last_pk == &self.public_keys[0] {
-            // The last public key closes a cycle
-            true
-        } else {
-            // A route with no repetitions
-            seen.insert(last_pk.clone())
-        }
-    }
-
-    /// Find two consecutive public keys (pk1, pk2) inside a friends route.
-    pub fn find_pk_pair(&self, pk1: &PublicKey, pk2: &PublicKey) -> Option<usize> {
-        let pks = &self.public_keys;
-        for i in 0 ..= pks.len().checked_sub(2)? {
-            if pk1 == &pks[i] && pk2 == &pks[i+1] {
-                return Some(i);
-            }
-        }
-        None
-    }
-
-    pub fn to_bytes(&self) -> Vec<u8> {
-        let mut res_bytes = Vec::new();
-        res_bytes.write_u64::<BigEndian>(
-            usize_to_u64(self.public_keys.len()).unwrap()).unwrap();
-        for public_key in &self.public_keys {
-            res_bytes.extend_from_slice(public_key);
-        }
-        res_bytes
-    }
-
-    /// Produce a cryptographic hash over the contents of the route.
-    pub fn hash(&self) -> HashResult {
-        hash::sha_512_256(&self.to_bytes())
-    }
-
-    /// Find the index of a public key inside the route.
-    /// source is considered to be index 0.
-    /// dest is considered to be the last index.
-    ///
-    /// Note that the returned index does not map directly to an 
-    /// index of self.route_links vector.
-    pub fn pk_to_index(&self, public_key: &PublicKey) -> Option<usize> {
-        self.public_keys
-            .iter()
-            .position(|cur_public_key| cur_public_key == public_key)
-    }
-
-    /// Get the public key of a node according to its index.
-    pub fn index_to_pk(&self, index: usize) -> Option<&PublicKey> {
-        self.public_keys.get(index)
-    }
-}
-
-
-impl Ratio<u128> {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut res_bytes = Vec::new();
-        match *self {
-            Ratio::One => {
-                res_bytes.write_u8(0).unwrap();
-            },
-            Ratio::Numerator(num) => {
-                res_bytes.write_u8(1).unwrap();
-                res_bytes.write_u128::<BigEndian>(num).unwrap();
-            },
-        }
-        res_bytes
-    }
-}
-
-impl FreezeLink {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut res_bytes = Vec::new();
-        res_bytes.write_u128::<BigEndian>(self.shared_credits)
-            .expect("Could not serialize u64!");
-        res_bytes.extend_from_slice(&self.usable_ratio.to_bytes());
-        res_bytes
-    }
-}
-
-impl RequestSendFunds {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut res_bytes = Vec::new();
-        res_bytes.extend_from_slice(&self.request_id);
-        res_bytes.extend_from_slice(&self.route.to_bytes());
-        res_bytes.write_u128::<BigEndian>(self.dest_payment).unwrap();
-        for freeze_link in &self.freeze_links {
-            res_bytes.extend_from_slice(&freeze_link.to_bytes());
-        }
-        res_bytes
-    }
-
-    pub fn create_pending_request(&self) -> PendingFriendRequest {
-        PendingFriendRequest {
-            request_id: self.request_id,
-            route: self.route.clone(),
-            dest_payment: self.dest_payment,
-            invoice_id: self.invoice_id.clone(),
-        }
-    }
-}
-
-impl ResponseSendFunds {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut res_bytes = Vec::new();
-        res_bytes.extend_from_slice(&self.request_id);
-        res_bytes.extend_from_slice(&self.rand_nonce);
-        res_bytes.extend_from_slice(&self.signature);
-        res_bytes
-    }
-}
-
-
-impl FailureSendFunds {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut res_bytes = Vec::new();
-        res_bytes.extend_from_slice(&self.request_id);
-        res_bytes.extend_from_slice(&self.reporting_public_key);
-        res_bytes.extend_from_slice(&self.signature);
-        res_bytes
-    }
-}
-
-#[allow(unused)]
-impl FriendTcOp {
-    pub fn to_bytes(&self) -> Vec<u8> {
-        let mut res_bytes = Vec::new();
-        match self {
-            FriendTcOp::EnableRequests => {
-                res_bytes.push(0u8);
-            },
-            FriendTcOp::DisableRequests => {
-                res_bytes.push(1u8);
-            }
-            FriendTcOp::SetRemoteMaxDebt(remote_max_debt) => {
-                res_bytes.push(2u8);
-                res_bytes.write_u128::<BigEndian>(*remote_max_debt)
-                    .expect("Failed to serialize u64 (remote_max_debt)");
-            }
-            FriendTcOp::RequestSendFunds(request_send_funds) => {
-                res_bytes.push(3u8);
-                res_bytes.append(&mut request_send_funds.to_bytes())
-            }
-            FriendTcOp::ResponseSendFunds(response_send_funds) => {
-                res_bytes.push(4u8);
-                res_bytes.append(&mut response_send_funds.to_bytes())
-            }
-            FriendTcOp::FailureSendFunds(failure_send_funds) => {
-                res_bytes.push(5u8);
-                res_bytes.append(&mut failure_send_funds.to_bytes())
-            }
-            
-        }
-        res_bytes
-    }
-
-    /// Get an approximation of the amount of bytes required to represent 
-    /// this operation.
-    pub fn approx_bytes_count(&self) -> usize {
-        self.to_bytes().len()
-    }
-}
 
 impl UserRequestSendFunds {
     pub fn to_request(self) -> RequestSendFunds {
@@ -548,24 +293,10 @@ impl SendFundsReceipt {
     }
 }
 
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct FriendMoveTokenRequest {
-    pub friend_move_token: FriendMoveToken,
-    // Do we want the remote side to return the token:
-    pub token_wanted: bool,
-}
-
 #[allow(unused)]
 pub struct FriendInconsistencyError {
     pub reset_token: Signature,
     pub balance_for_reset: i128,
-}
-
-
-#[derive(Debug, Clone)]
-pub enum FriendMessage {
-    MoveTokenRequest(FriendMoveTokenRequest),
-    InconsistencyError(ResetTerms),
 }
 
 
@@ -619,11 +350,4 @@ pub enum FunderOutgoingControl<A: Clone> {
 pub enum FunderOutgoingComm<A> {
     FriendMessage((PublicKey, FriendMessage)),
     ChannelerConfig(ChannelerConfig<A>),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ResetTerms {
-    pub reset_token: Signature,
-    pub inconsistency_counter: u64,
-    pub balance_for_reset: i128,
 }
