+++ components/proto/Cargo.toml
@@ -7,9 +7,15 @@ build = "build.rs"
 [dependencies]
 
 offst-crypto = { path = "../crypto", version = "0.1.0" }
+offst-utils = { path = "../utils", version = "0.1.0"}
 
 capnp = "0.8"
 byteorder = "1.1"
+
+serde = "1"
+serde_derive = "1"
+serde_json = "1.0.27"
+bytes = "0.4"
 
 [dev-dependencies]
