+++ components/proto/src/funder/mod.rs
@@ -1,1 +1,1 @@
-mod messages;
+pub mod messages;
