+++ components/proto/src/lib.rs
@@ -4,13 +4,22 @@
 extern crate capnp;
 extern crate byteorder;
 
+#[macro_use]
+extern crate offst_utils as utils;
 extern crate offst_crypto as crypto;
+
+extern crate serde;
+#[macro_use]
+extern crate serde_derive;
+extern crate bytes;
 
 #[macro_use]
 pub mod macros;
 pub mod capnp_custom_int;
 pub mod relay;
 pub mod secure_channel;
+pub mod funder;
+
 
 
 include_schema!(channeler_capnp, "channeler_capnp");
